pyproject.toml
[tool.poetry]
name = "attey"
version = "0.1.0"
description = "Manager bot for the @rcade Discord project."
authors = ["Adriene Kae <polyjitter@fastmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
hikari = "^2.0.0-alpha.109"
hikari-miru = "^1.1.1"
hikari-clusters = "^0.2.5"
hikari-lightbulb = "^2.2.2"
rethinkdb = "^2.4.9"
uvloop = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
